move quick_sort to compatibility
